Make callbacks generic.
Move the wrapper around a method which calls a coderef supplied to the
constructor (possibly as a string which is eval'd) into a parameterized role.

Write a small set of traits for the JobRunner which used it allowing you to
wrap things trivially (for connecting to the DB before running a job or
whatever). This will get fugly fast if you try and put a real amount of code
in here, but for oneliners it's entirely fine